Dedupe types
